fix: change EanParseResult fields from Double to BigDecimal to match parser types

+++ backend/src/main/java/com/acougue/modules/balanca/dto/EanParseResult.java
@@ -4,11 +4,13 @@ import com.acougue.modules.balanca.EanBalancaParser.FormatoEtiqueta;
 import lombok.Builder;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Value
 @Builder
 public class EanParseResult {
     int codigoBalanca;
-    Double valorTotal;
-    double pesoKg;
+    BigDecimal valorTotal;
+    BigDecimal pesoKg;
     FormatoEtiqueta formato;
 }
